Update import path, fix to work with multiple migrations.
The migrations we were running were all using the same table to track
their migration status. This led to some confusion and made it
impossible to apply migrations for all the projects.

By refactoring, we can use a table for each project, which is more
tables than I'd like but is probably cleaner.

We also still were using the old package name in go.mod, so that got
updated.